Moved jsonapi types
