+++ sciety_labs/app/routers/api/categorisation/typing.py
@@ -2,17 +2,6 @@ from typing import Optional, Sequence
 from typing_extensions import NotRequired, TypedDict
 
 
-class JsonApiErrorDict(TypedDict):
-    # https://jsonapi.org/format/#errors
-    status: NotRequired[str]
-    title: NotRequired[str]
-    detail: NotRequired[str]
-
-
-class JsonApiErrorsResponseDict(TypedDict):
-    errors: Sequence[JsonApiErrorDict]
-
-
 class CategorisationDict(TypedDict):
     display_name: str
     type: str
