+++ sciety_labs/app/routers/api/utils/jsonapi_typing.py
@@ -1,0 +1,13 @@
+from typing import Sequence
+from typing_extensions import NotRequired, TypedDict
+
+
+class JsonApiErrorDict(TypedDict):
+    # https://jsonapi.org/format/#errors
+    status: NotRequired[str]
+    title: NotRequired[str]
+    detail: NotRequired[str]
+
+
+class JsonApiErrorsResponseDict(TypedDict):
+    errors: Sequence[JsonApiErrorDict]
